Fixing `recommendationcategory` naming
Now made being written as `recommendation_category`. Will push a matching update to the actual API, along with change to `marketplaceid`.

This is a bit of a break from current codebase, but we are prepping for 1.0 release anyway, so we can document the breaking changes as we move forward. I feel it's better to be correct in this instance, rather than keep it backwards-compatible.