Remove double url encoding for report options (#247)
* Remove double url encoding for report options

Using `urllib.parse.quote` had good intentions, I believe, matching
the expected output MWS stated in their docs. However, the same quote
action is performed elsewhere in the chain, for all data sent through
a request, and before that data is joined to keys in the body of the
request.

So, performing that quote action here results in double encoding,
turning `=` into `%3D`, then again into `%253D` (encoding `%`).
This resulted in broken requests.

Test added to ensure this usage stays in place in future.

* Adding ubuntu 20.04 to dev test suite matrix

* Readme encoding to utf-8 for setup.py

Otherwise, it seems it can't properly read the fancy emojis and such.