Proper POST request formation (#253)
* Align Feeds request method code style with other request methods

Using the `data` variable explicitly, then passing it;
instead of collapsing it to this one-liner (that was actually longer).

* Add enums to Feeds API.

- FeedProcessingStatus
- FeedType
- Both new enums attached to Feeds class for convenient access
- Docs updated

* Incorrect docstring for CurrencyCode.

* Implement correct POST data handling

- Defaults all request methods to POST
- Removes method="POST" from other request methods.
- Attaches request params to the body of the request thorugh `data`
  instead of manually adding them to the request URL, as done prior.

An exception had to be made for the `body` kwarg in `make_request`.
This was mainly used by `Feeds.submit_feed` to add the feed content
file to the body of the request, and that still needs to function as
it did before. In that case only, params are added to `params` in the
request, which pushes them into the URL query string.

This change led to a number of other changes, hence the diff on this
commit is a bit large. The main cause was parameter cleaning and how
it related to:

1. the canonicalized query string; and
2. the body of the request itself.

Originally, all params were URL encoded first. This is required for the
canonical query string and request signing, and it's also a requirement
for using params in the URL itself. So, that used to be easy.

Now that data is moved into the POST body, MWS expects the *not*-encoded
values for that data, else they don't accept it (malformed data).
But, the canonical string we sign still needs to be URL-encoded first,
otherwise Amazon's calculated signature is different.

This was solved by splitting the cleaning logic within
`clean_params_dict` with an optional flag, `urlencode` (default False).
That way, request parameters could be cleaned without being encoded,
so they would have appropriate string values when added to POST body.
Meanwhile, `canonicalized_query_string` (a sensible rename of the
`request_description` function) makes use of this flag to get encoded
values, producing a string in the required format.

Finally, we remove the kludgy logic for building a request URL,
originally done by manually adding the request params to the URL string.
Now, we just use `request`'s tooling, either `params` or `data` args
as needed.

(Testing shows Amazon WILL NOT accept GET requests with a body,
so there's some extra checks in place to ensure params get to the
right place).